Errors fixed

+++ homework_guess/random_value.cpp
@@ -1,11 +1,10 @@
 #include <iostream>
-#include "random_value.h"
 #include <cstdlib>
 #include <ctime>
 
-int getRandomValue() {
-
-	std::srand(std::time(nullptr)); 
+int getRandomValue() 
+{
+	std::srand((unsigned int)std::time(nullptr)); 
      
 	const int max_value = 100;
 	const int random_value = std::rand() % 100;
